requirements_test.txt
django-cms
coverage
django-coverage
factory_boy
mock
selenium
tox